+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/CosmosItemTests.cs
@@ -65,7 +65,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task CreateDropItemTest()
         {
-            ToDoActivity testItem = this.CreateRandomToDoActivity();
+            ToDoActivity testItem = ToDoActivity.CreateRandomToDoActivity();
             ItemResponse<ToDoActivity> response = await this.Container.CreateItemAsync<ToDoActivity>(item: testItem);
             Assert.IsNotNull(response);
             Assert.IsNotNull(response.MaxResourceQuota);
@@ -318,7 +318,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task CreateDropItemStreamTest()
         {
-            ToDoActivity testItem = this.CreateRandomToDoActivity();
+            ToDoActivity testItem = ToDoActivity.CreateRandomToDoActivity();
             using (Stream stream = TestCommon.Serializer.ToStream<ToDoActivity>(testItem))
             {
                 using (ResponseMessage response = await this.Container.CreateItemStreamAsync(partitionKey: new Cosmos.PartitionKey(testItem.status), streamPayload: stream))
@@ -343,7 +343,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task UpsertItemStreamTest()
         {
-            ToDoActivity testItem = this.CreateRandomToDoActivity();
+            ToDoActivity testItem = ToDoActivity.CreateRandomToDoActivity();
             using (Stream stream = TestCommon.Serializer.ToStream<ToDoActivity>(testItem))
             {
                 //Create the object
@@ -378,7 +378,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task ReplaceItemStreamTest()
         {
-            ToDoActivity testItem = this.CreateRandomToDoActivity();
+            ToDoActivity testItem = ToDoActivity.CreateRandomToDoActivity();
             using (Stream stream = TestCommon.Serializer.ToStream<ToDoActivity>(testItem))
             {
                 //Replace a non-existing item. It should fail, and not throw an exception.
@@ -426,7 +426,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [DataTestMethod]
         public async Task ItemStreamIterator(bool useStatelessIterator)
         {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(3, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, 3, randomPartitionKey: true);
             HashSet<string> itemIds = deleteList.Select(x => x.id).ToHashSet<string>();
 
             string lastContinuationToken = null;
@@ -464,55 +464,6 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
 
                 }
 
-            }
-
-            Assert.IsNull(lastContinuationToken);
-            Assert.AreEqual(itemIds.Count, 0);
-        }
-
-        [DataRow(false)]
-        [DataRow(true)]
-        [DataTestMethod]
-        public async Task ItemLinqReadFeedTest(bool useStatelessIterator)
-        {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(3, randomPartitionKey: true);
-            HashSet<string> itemIds = deleteList.Select(x => x.id).ToHashSet<string>();
-
-            QueryRequestOptions requestOptions = new QueryRequestOptions()
-            {
-                MaxItemCount = 1
-            };
-
-            List<ToDoActivity> itemsViaReadFeed = this.Container.GetItemLinqQueryable<ToDoActivity>(
-                allowSynchronousQueryExecution: true,
-                requestOptions: requestOptions).ToList();
-
-            Assert.IsTrue(itemsViaReadFeed.Count >= 3);
-            CollectionAssert.AreEqual(deleteList.ToList(), itemsViaReadFeed);
-
-            string lastContinuationToken = null;
-            FeedIterator<ToDoActivity> feedIterator = this.Container.GetItemLinqQueryable<ToDoActivity>(
-                requestOptions: requestOptions).ToFeedIterator();
-
-            while (feedIterator.HasMoreResults)
-            {
-                if (useStatelessIterator)
-                {
-                    feedIterator = this.Container.GetItemLinqQueryable<ToDoActivity>(
-                        continuationToken: lastContinuationToken,
-                        requestOptions: requestOptions).ToFeedIterator();
-                }
-
-                var responseMessage = await feedIterator.ReadNextAsync(this.cancellationToken);
-                lastContinuationToken = responseMessage.ContinuationToken;
-
-                foreach (ToDoActivity toDoActivity in responseMessage)
-                {
-                    if (itemIds.Contains(toDoActivity.id))
-                    {
-                        itemIds.Remove(toDoActivity.id);
-                    }
-                }
             }
 
             Assert.IsNull(lastContinuationToken);
@@ -651,7 +602,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task ItemIterator()
         {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(3, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, 3, randomPartitionKey: true);
             HashSet<string> itemIds = deleteList.Select(x => x.id).ToHashSet<string>();
             FeedIterator<ToDoActivity> feedIterator =
                 this.Container.GetItemQueryIterator<ToDoActivity>();
@@ -781,7 +732,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [DataTestMethod]
         public async Task QuerySinglePartitionItemStreamTest(int perPKItemCount, int maxItemCount)
         {
-            IList<ToDoActivity> deleteList = deleteList = await this.CreateRandomItems(pkCount: 3, perPKItemCount: perPKItemCount, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = deleteList = await ToDoActivity.CreateRandomItems(this.Container, pkCount: 3, perPKItemCount: perPKItemCount, randomPartitionKey: true);
             ToDoActivity find = deleteList.First();
 
             QueryDefinition sql = new QueryDefinition("select * from r where r.status = @status").WithParameter("@status", find.status);
@@ -850,7 +801,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task ItemMultiplePartitionQuery()
         {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(3, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, 3, randomPartitionKey: true);
 
             ToDoActivity find = deleteList.First();
             QueryDefinition sql = new QueryDefinition("select * from toDoActivity t where t.id = '" + find.id + "'");
@@ -890,7 +841,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 System.Globalization.CultureInfo.GetCultureInfo("fr-FR")
             };
 
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(300, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, 300, randomPartitionKey: true);
 
             try
             {
@@ -950,7 +901,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task ItemMultiplePartitionQueryStream()
         {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(101, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, 101, randomPartitionKey: true);
             QueryDefinition sql = new QueryDefinition("SELECT * FROM toDoActivity t");
 
             QueryRequestOptions requestOptions = new QueryRequestOptions()
@@ -992,10 +943,10 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         public async Task ItemSinglePartitionQueryStream()
         {
             //Create a 101 random items with random guid PK values
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(pkCount: 101, perPKItemCount: 1, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, pkCount: 101, perPKItemCount: 1, randomPartitionKey: true);
 
             // Create 10 items with same pk value
-            IList<ToDoActivity> findItems = await this.CreateRandomItems(pkCount: 1, perPKItemCount: 10, randomPartitionKey: false);
+            IList<ToDoActivity> findItems = await ToDoActivity.CreateRandomItems(this.Container, pkCount: 1, perPKItemCount: 10, randomPartitionKey: false);
 
             string findPkValue = findItems.First().status;
             QueryDefinition sql = new QueryDefinition("SELECT * FROM toDoActivity t where t.status = @pkValue").WithParameter("@pkValue", findPkValue);
@@ -1130,7 +1081,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task ItemQueryStreamSerializationSetting()
         {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(101, randomPartitionKey: true);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(this.Container, 101, randomPartitionKey: true);
 
             QueryDefinition sql = new QueryDefinition("SELECT * FROM toDoActivity t ORDER BY t.taskNum");
             CosmosSerializationOptions options = new CosmosSerializationOptions(
@@ -1196,7 +1147,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task ValidateMaxItemCountOnItemQuery()
         {
-            IList<ToDoActivity> deleteList = await this.CreateRandomItems(pkCount: 1, perPKItemCount: 6, randomPartitionKey: false);
+            IList<ToDoActivity> deleteList = await ToDoActivity.CreateRandomItems(container: this.Container, pkCount: 1, perPKItemCount: 6, randomPartitionKey: false);
 
             ToDoActivity toDoActivity = deleteList.First();
             QueryDefinition sql = new QueryDefinition(
@@ -1241,7 +1192,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task NegativeQueryTest()
         {
-            IList<ToDoActivity> items = await this.CreateRandomItems(pkCount: 10, perPKItemCount: 20, randomPartitionKey: true);
+            IList<ToDoActivity> items = await ToDoActivity.CreateRandomItems(container: this.Container, pkCount: 10, perPKItemCount: 20, randomPartitionKey: true);
 
             try
             {
@@ -1276,7 +1227,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         public async Task ItemRequestOptionAccessConditionTest()
         {
             // Create an item
-            ToDoActivity testItem = (await this.CreateRandomItems(1, randomPartitionKey: true)).First();
+            ToDoActivity testItem = (await ToDoActivity.CreateRandomItems(this.Container, 1, randomPartitionKey: true)).First();
 
             ItemRequestOptions itemRequestOptions = new ItemRequestOptions()
             {
@@ -1330,7 +1281,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 Assert.AreEqual(nonPartitionItemId, response.Resource.id);
 
                 //Adding item to fixed container with CosmosContainerSettings.NonePartitionKeyValue.
-                ToDoActivity itemWithoutPK = this.CreateRandomToDoActivity();
+                ToDoActivity itemWithoutPK = ToDoActivity.CreateRandomToDoActivity();
                 ItemResponse<ToDoActivity> createResponseWithoutPk = await fixedContainer.CreateItemAsync<ToDoActivity>(
                  item: itemWithoutPK,
                  partitionKey: Cosmos.PartitionKey.None);
@@ -1430,130 +1381,6 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                     await fixedContainer.DeleteContainerAsync();
                 }
             }
-        }
-
-        [TestMethod]
-        public async Task ItemLINQQueryTest()
-        {
-            //Creating items for query.
-            IList<ToDoActivity> itemList = await this.CreateRandomItems(pkCount: 2, perPKItemCount: 1, randomPartitionKey: true);
-
-            IOrderedQueryable<ToDoActivity> linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>();
-            IQueryable<ToDoActivity> queriable = linqQueryable.Where(item => (item.taskNum < 100));
-            //V3 Asynchronous query execution with LINQ query generation sql text.
-            FeedIterator<ToDoActivity> setIterator = this.Container.GetItemQueryIterator<ToDoActivity>(
-                queriable.ToQueryDefinition(),
-                requestOptions: new QueryRequestOptions() { MaxConcurrency = 2 });
-
-            int resultsFetched = 0;
-            while (setIterator.HasMoreResults)
-            {
-                FeedResponse<ToDoActivity> queryResponse = await setIterator.ReadNextAsync();
-                resultsFetched += queryResponse.Count();
-
-                // For the items returned with NonePartitionKeyValue
-                IEnumerator<ToDoActivity> iter = queryResponse.GetEnumerator();
-                while (iter.MoveNext())
-                {
-                    ToDoActivity activity = iter.Current;
-                    Assert.AreEqual(42, activity.taskNum);
-                }
-                Assert.AreEqual(2, resultsFetched);
-            }
-
-            //Checking for exception in case of ToFeedIterator() use on non cosmos linq IQueryable.
-            try
-            {
-                IQueryable<ToDoActivity> nonLinqQueryable = (new List<ToDoActivity> { this.CreateRandomToDoActivity() }).AsQueryable();
-                setIterator = nonLinqQueryable.ToFeedIterator();
-                Assert.Fail("It should throw ArgumentOutOfRangeException as ToFeedIterator() only applicable to cosmos LINQ query");
-            }
-            catch (ArgumentOutOfRangeException ex)
-            {
-                Assert.IsTrue(ex.Message.Contains("ToFeedIterator is only supported on cosmos LINQ query operations"));
-            }
-
-            //LINQ query execution without partition key.
-            linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(allowSynchronousQueryExecution: true);
-            queriable = linqQueryable.Where(item => (item.taskNum < 100));
-
-            Assert.AreEqual(2, queriable.Count());
-            Assert.AreEqual(itemList[0].id, queriable.ToList()[0].id);
-            Assert.AreEqual(itemList[1].id, queriable.ToList()[1].id);
-
-            //LINQ query execution with wrong partition key.
-            linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(
-                allowSynchronousQueryExecution: true,
-                requestOptions: new QueryRequestOptions() { PartitionKey = new Cosmos.PartitionKey("test") });
-            queriable = linqQueryable.Where(item => (item.taskNum < 100));
-            Assert.AreEqual(0, queriable.Count());
-
-            //LINQ query execution with correct partition key.
-            linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(
-                allowSynchronousQueryExecution: true,
-                requestOptions: new QueryRequestOptions { ConsistencyLevel = Cosmos.ConsistencyLevel.Eventual, PartitionKey = new Cosmos.PartitionKey(itemList[1].status) });
-            queriable = linqQueryable.Where(item => (item.taskNum < 100));
-            Assert.AreEqual(1, queriable.Count());
-            Assert.AreEqual(itemList[1].id, queriable.ToList()[0].id);
-
-            //Creating LINQ query without setting allowSynchronousQueryExecution true.
-            try
-            {
-                linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(
-                     requestOptions: new QueryRequestOptions() { PartitionKey = new Cosmos.PartitionKey(itemList[0].status) });
-                linqQueryable.Where(item => (item.taskNum < 100)).ToList();
-                Assert.Fail("Should throw NotSupportedException");
-            }
-            catch (NotSupportedException exception)
-            {
-                Assert.IsTrue(exception.Message.Contains("To execute LINQ query please set allowSynchronousQueryExecution true"));
-            }
-        }
-
-        [TestMethod]
-        public async Task ItemLINQQueryWithContinuationTokenTest()
-        {
-            //Creating items for query.
-            IList<ToDoActivity> itemList = await this.CreateRandomItems(pkCount: 10, perPKItemCount: 1, randomPartitionKey: true);
-
-            QueryRequestOptions queryRequestOptions = new QueryRequestOptions();
-            queryRequestOptions.MaxConcurrency = 1;
-            queryRequestOptions.MaxItemCount = 5;
-            IOrderedQueryable<ToDoActivity> linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(requestOptions: queryRequestOptions);
-            IQueryable<ToDoActivity> queriable = linqQueryable.Where(item => (item.taskNum < 100));
-            FeedIterator<ToDoActivity> feedIterator = queriable.ToFeedIterator();
-
-            int firstItemSet = 0;
-            string continuationToken = null;
-            while (feedIterator.HasMoreResults)
-            {
-                FeedResponse<ToDoActivity> feedResponse = await feedIterator.ReadNextAsync();
-                firstItemSet = feedResponse.Count();
-                continuationToken = feedResponse.ContinuationToken;
-                if (firstItemSet > 0)
-                {
-                    break;
-                }
-            }
-
-            linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(continuationToken: continuationToken, requestOptions: queryRequestOptions);
-            queriable = linqQueryable.Where(item => (item.taskNum < 100));
-            feedIterator = queriable.ToFeedIterator();
-
-            //Test continuationToken with LINQ query generation and asynchronous feedIterator execution.
-            int secondItemSet = 0;
-            while (feedIterator.HasMoreResults)
-            {
-                FeedResponse<ToDoActivity> feedResponse = await feedIterator.ReadNextAsync();
-                secondItemSet += feedResponse.Count();
-            }
-
-            Assert.AreEqual(10 - firstItemSet, secondItemSet);
-
-            //Test continuationToken with blocking LINQ execution
-            linqQueryable = this.Container.GetItemLinqQueryable<ToDoActivity>(allowSynchronousQueryExecution: true, continuationToken: continuationToken, requestOptions: queryRequestOptions);
-            int linqExecutionItemCount = linqQueryable.Where(item => (item.taskNum < 100)).Count();
-            Assert.AreEqual(10 - firstItemSet, linqExecutionItemCount);
         }
 
         // Move the data from None Partition to other logical partitions
@@ -1658,7 +1485,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 Task[] createItems = new Task[300];
                 for (int i = 0; i < createItems.Length; i++)
                 {
-                    ToDoActivity temp = this.CreateRandomToDoActivity();
+                    ToDoActivity temp = ToDoActivity.CreateRandomToDoActivity();
                     createItems[i] = container.CreateItemStreamAsync(
                         partitionKey: new Cosmos.PartitionKey(temp.status),
                         streamPayload: TestCommon.Serializer.ToStream<ToDoActivity>(temp));
@@ -1693,7 +1520,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
         [TestMethod]
         public async Task VerifySessionTokenPassThrough()
         {
-            ToDoActivity temp = this.CreateRandomToDoActivity("TBD");
+            ToDoActivity temp = ToDoActivity.CreateRandomToDoActivity("TBD");
 
             ItemResponse<ToDoActivity> responseAstype = await this.Container.CreateItemAsync<ToDoActivity>(partitionKey: new Cosmos.PartitionKey(temp.status), item: temp);
 
@@ -1818,32 +1645,6 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
             }
         }
 
-        private async Task<IList<ToDoActivity>> CreateRandomItems(int pkCount, int perPKItemCount = 1, bool randomPartitionKey = true)
-        {
-            Assert.IsFalse(!randomPartitionKey && pkCount > 1);
-
-            List<ToDoActivity> createdList = new List<ToDoActivity>();
-            for (int i = 0; i < pkCount; i++)
-            {
-                string pk = "TBD";
-                if (randomPartitionKey)
-                {
-                    pk += Guid.NewGuid().ToString();
-                }
-
-                for (int j = 0; j < perPKItemCount; j++)
-                {
-                    ToDoActivity temp = this.CreateRandomToDoActivity(pk);
-
-                    createdList.Add(temp);
-
-                    await this.Container.CreateItemAsync<ToDoActivity>(item: temp);
-                }
-            }
-
-            return createdList;
-        }
-
         private async Task<ContainerCore> CreateNonPartitionedContainer(string id)
         {
             await CosmosItemTests.CreateNonPartitionedContainer(
@@ -1904,7 +1705,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
             client.DefaultRequestHeaders.Add("x-ms-version", CosmosItemTests.PreNonPartitionedMigrationApiVersion);
             client.DefaultRequestHeaders.Add("authorization", authHeader);
 
-            string itemDefinition = JsonConvert.SerializeObject(this.CreateRandomToDoActivity(id: itemId));
+            string itemDefinition = JsonConvert.SerializeObject(ToDoActivity.CreateRandomToDoActivity(id: itemId));
             {
                 StringContent itemContent = new StringContent(itemDefinition);
                 Uri requestUri = new Uri(baseUri, resourceLink);
@@ -1960,57 +1761,6 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 keyType,
                 tokenVersion,
                 signature));
-        }
-
-        private ToDoActivity CreateRandomToDoActivity(string pk = null, string id = null)
-        {
-            if (string.IsNullOrEmpty(pk))
-            {
-                pk = "TBD" + Guid.NewGuid().ToString();
-            }
-            if (id == null)
-            {
-                id = Guid.NewGuid().ToString();
-            }
-            return new ToDoActivity()
-            {
-                id = id,
-                description = "CreateRandomToDoActivity",
-                status = pk,
-                taskNum = 42,
-                cost = double.MaxValue
-            };
-        }
-
-
-
-        public class ToDoActivity
-        {
-            public string id { get; set; }
-            public int taskNum { get; set; }
-            public double cost { get; set; }
-            public string description { get; set; }
-            public string status { get; set; }
-
-            public override bool Equals(Object obj)
-            {
-                ToDoActivity input = obj as ToDoActivity;
-                if (input == null)
-                {
-                    return false;
-                }
-
-                return string.Equals(this.id, input.id)
-                    && this.taskNum == input.taskNum
-                    && this.cost == input.cost
-                    && string.Equals(this.description, input.description)
-                    && string.Equals(this.status, input.status);
-            }
-
-            public override int GetHashCode()
-            {
-                return base.GetHashCode();
-            }
         }
 
         public class ToDoActivityAfterMigration